+++ src/gort/overwatcher/alerts.py
@@ -9,10 +9,12 @@
 from __future__ import annotations
 
 import asyncio
+import enum
 from time import time
 
 from typing import TYPE_CHECKING
 
+from lvmopstools.retrier import Retrier
 from pydantic import BaseModel
 
 from gort.overwatcher.core import OverwatcherModuleTask
@@ -24,7 +26,7 @@ if TYPE_CHECKING:
     pass
 
 
-__all__ = ["AlertsOverwatcher"]
+__all__ = ["AlertsOverwatcher", "ActiveAlert"]
 
 
 class AlertsSummary(BaseModel):
@@ -41,6 +43,19 @@ class AlertsSummary(BaseModel):
     o2_room_alerts: dict[str, bool] | None = None
     heater_alert: bool | None = None
     heater_camera_alerts: dict[str, bool] | None = None
+
+
+class ActiveAlert(enum.Flag):
+    """Flags for active alerts."""
+
+    HUMIDITY = enum.auto()
+    DEW_POINT = enum.auto()
+    WIND = enum.auto()
+    RAIN = enum.auto()
+    DOOR = enum.auto()
+    CAMERA_TEMPERATURE = enum.auto()
+    O2 = enum.auto()
+    UNKNOWN = enum.auto()
 
 
 class AlertsMonitorTask(OverwatcherModuleTask["AlertsOverwatcher"]):
@@ -96,6 +111,8 @@ class AlertsMonitorTask(OverwatcherModuleTask["AlertsOverwatcher"]):
         if data is None:
             raise ValueError("no alerts data available.")
 
+        # For some very critical alerts, we require them to be not null (null here
+        # means no data was available or the API failed getting the alert data).
         if data.rain is None or data.humidity_alert is None or data.wind_alert is None:
             raise ValueError("incomplete alerts data.")
 
@@ -115,36 +132,58 @@ class AlertsOverwatcher(OverwatcherModule):
         self.state: AlertsSummary | None = None
         self.locked_until: float = 0
 
-    def is_safe(self):
+    def is_safe(self) -> tuple[bool, ActiveAlert]:
         """Determines whether it is safe to open."""
+
         if self.state is None:
             self.log.warning("Alerts data not available. is_safe() returns False.")
-            return False
+            return False, ActiveAlert.UNKNOWN
 
         # If we have issued a previous unsafe alert, the main task will close the dome
         # and put a lock for 30 minutes to prevent the dome from opening/closing too
         # frequently if the weather is unstable.
         if self.locked_until > 0 and time() < self.locked_until:
-            return False
+            return False, ActiveAlert(0)
+
+        is_safe: bool = True
+        active_alerts = ActiveAlert(0)
 
         if self.state.rain:
             self.log.warning("Rain alert detected.")
-            return False
-        elif self.state.humidity_alert:
+            active_alerts |= ActiveAlert.RAIN
+            is_safe = False
+        if self.state.humidity_alert:
             self.log.warning("Humidity alert detected.")
-            return False
-        elif self.state.dew_point_alert:
+            active_alerts |= ActiveAlert.HUMIDITY
+            is_safe = False
+        if self.state.dew_point_alert:
             self.log.warning("Dew point alert detected.")
-            return False
-        elif self.state.wind_alert:
+            active_alerts |= ActiveAlert.DEW_POINT
+            is_safe = False
+        if self.state.wind_alert:
             self.log.warning("Wind alert detected.")
-            return False
-        else:
-            self.locked_until = 0
-            return True
+            active_alerts |= ActiveAlert.WIND
+            is_safe = False
 
-    @staticmethod
-    async def get_alerts_summary() -> AlertsSummary:
+        # These alerts are not critical but we log them.
+        # TODO: maybe we do want to do something about these alerts.
+        if self.state.door_alert:
+            self.log.warning("Door alert detected.")
+            active_alerts |= ActiveAlert.DOOR
+        if self.state.camera_temperature_alert:
+            self.log.warning("Camera temperature alert detected.")
+            active_alerts |= ActiveAlert.CAMERA_TEMPERATURE
+        if self.state.o2_alert:
+            self.log.warning("O2 alert detected.")
+            active_alerts |= ActiveAlert.O2
+
+        if is_safe:
+            self.locked_until = 0
+
+        return is_safe, active_alerts
+
+    @Retrier(max_attempts=3, delay=5)
+    async def get_alerts_summary(self) -> AlertsSummary:
         """Returns the alerts report."""
 
         data = await get_lvmapi_route("/alerts/summary")