+++ src/gort/overwatcher/overwatcher.py
@@ -76,7 +76,7 @@ class OverwatcherMainTask(OverwatcherTask):
             self.previous_state = copy(ow.state)
 
             try:
-                is_safe = ow.alerts.is_safe()
+                is_safe, _ = ow.alerts.is_safe()
                 is_night = ow.ephemeris.is_night()
 
                 ow.state.night = is_night
@@ -127,7 +127,7 @@ class OverwatcherMainTask(OverwatcherTask):
 
         if not closed or observing or calibrating:
             await self.overwatcher.notify(
-                "Closing the dome due to unsafe conditions.",
+                "Unsafe conditions detected.",
                 level="warning",
             )
 
@@ -151,6 +151,7 @@ class OverwatcherMainTask(OverwatcherTask):
                 await self.overwatcher.calibrations.cancel()
 
             if not closed:
+                await self.overwatcher.notify("Closing the dome.")
                 await self.overwatcher.dome.shutdown(retry=True, park=True)
 
                 # If we have to close because of unsafe conditions, we don't want
@@ -273,6 +274,7 @@ class Overwatcher(NotifierMixIn):
             EphemerisOverwatcher,
             EventsOverwatcher,
             ObserverOverwatcher,
+            SafetyOverwatcher,
         )
 
         # Check if the instance already exists, in which case do nothing.
@@ -296,6 +298,7 @@ class Overwatcher(NotifierMixIn):
         # A series of tasks that must be run once every day.
         self.daily_tasks = DailyTasks(self)
 
+        self.safe = SafetyOverwatcher(self)
         self.ephemeris = EphemerisOverwatcher(self)
         self.calibrations = CalibrationsOverwatcher(self, calibrations_file)
         self.observer = ObserverOverwatcher(self)